Corrected language IDs
Language IDs have been corrected as follows: For each concept we computed
the set of language IDs present and compared this to the set of all valid
language IDs. If one additional ID was present and one valid ID missing,
the additional ID was changed to the missing one.
In a few cases more than one valid ID was missing. In these cases, the
additional ID was changed to the missing one with which it shared one letter.