Refactored glottolog access
Glottolog languoid information is now available as attribute
`glottolog_languoids` on the `dataset` object when a dataset's
cldf function is called.

This allows for in particular easy lookup of macroareas after
CLDF creation, thereby fixing #72.

This PR is also part of a bigger effort to extract statistics on
CLDF datasets right after creation (i.e. in the `__exit__` method
of a pylexibank CLDF dataset).